+++ limbus-staging/src/main/java/com/remondis/limbus/staging/LimbusStage.java
@@ -1,29 +1,21 @@
 package com.remondis.limbus.staging;
 
-import static java.util.Objects.nonNull;
+import static java.util.Objects.requireNonNull;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.security.Permission;
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.Semaphore;
 
-import org.jboss.shrinkwrap.api.exporter.ZipExporter;
+import org.jboss.shrinkwrap.api.formatter.Formatters;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
+import org.jboss.shrinkwrap.resolver.api.maven.Maven;
 
-import com.remondis.limbus.api.Classpath;
 import com.remondis.limbus.api.IInitializable;
 import com.remondis.limbus.api.LimbusException;
 import com.remondis.limbus.engine.api.LimbusEngine;
 import com.remondis.limbus.launcher.EngineLauncher;
-import com.remondis.limbus.staging.staging.Handler;
 import com.remondis.limbus.system.LimbusSystem;
-import com.remondis.limbus.utils.Lang;
 
 /**
  * The {@link LimbusStage} is a configuration holder for a Limbus environment that is ready to get bootstrapped. The
@@ -39,18 +31,13 @@ public final class LimbusStage {
 
   private boolean hasAccess = false;
 
-  private JavaArchive deployment;
-  private List<JavaArchive> dependencies;
   private LimbusSystem system;
 
-  private String deployName;
-  private Set<Permission> permissions;
+  private List<JavaArchive> dependencies;
 
-  private Classpath classpath;
-
-  LimbusStage() {
+  LimbusStage(LimbusSystem limbusSystem) {
+    this.system = limbusSystem;
     this.dependencies = new LinkedList<>();
-    this.permissions = new HashSet<>();
   }
 
   /**
@@ -66,31 +53,10 @@ public final class LimbusStage {
     EngineLauncher.skipSystemExit = true;
     try {
       EngineLauncher.bootstrapLimbusSystem(system);
-      deploy();
     } catch (Exception e) {
       stopStage();
       throw e;
     }
-  }
-
-  /**
-   * @return Returns the deployName of the resulting plugin deployment.
-   */
-  public String getDeployName() {
-    return deployName;
-  }
-
-  /**
-   * Returns the classpath that is available after starting this {@link LimbusStage}.
-   *
-   * <p>
-   * <b>Only available after starting the {@link LimbusStage} using {@link #startStage()}.
-   * </p>
-   *
-   * @return Returns the classpath of the deployment.
-   */
-  public Classpath getClasspath() {
-    return classpath;
   }
 
   /**
@@ -126,59 +92,10 @@ public final class LimbusStage {
     return system.hasComponent(requestType);
   }
 
-  private void deploy() throws LimbusException, IOException {
-    if (nonNull(deployment)) {
-      LimbusEngine limbusEngine = system.getComponent(LimbusEngine.class);
-      List<URL> urls = new LinkedList<>();
-      List<JavaArchive> archives = new LinkedList<>(dependencies);
-      for (JavaArchive a : archives) {
-        addJavaArchiveAsURL(urls, a, null);
-      }
-      addJavaArchiveAsURL(urls, deployment, null);
-      this.classpath = Classpath.create(deployName)
-          .add(urls);
-      limbusEngine.deployPlugin(classpath, permissions);
-    }
-  }
-
-  private void addJavaArchiveAsURL(List<URL> urls, JavaArchive a, String archiveName)
-      throws IOException, MalformedURLException {
-    byte[] resource = archiveToBytes(a);
-    String jarURLStr = null;
-    if (Lang.isEmpty(archiveName)) {
-      jarURLStr = createJarURL(a);
-    } else {
-      jarURLStr = createJarURL(archiveName);
-    }
-    URL jarURL = new URL(jarURLStr);// , resourceHandler);
-    urls.add(jarURL);
-    addResourceToStagingHandler(resource, jarURL);
-  }
-
-  private void addResourceToStagingHandler(byte[] resource, URL jarURL) {
-    if (Handler.CURRENT_INSTANCE == null) {
-      throw new IllegalStateException(
-          "The staging resource URL handler is not available. Call LimbusStaging.prepareEnvironment() before using LimbusStage.");
-    } else {
-      Handler.CURRENT_INSTANCE.addResource(jarURL, resource);
-    }
-  }
-
-  private String createJarURL(String archiveName) {
-    return String.format("staging:/%s.jar", archiveName);
-  }
-
-  private String createJarURL(JavaArchive jar) {
-    return String.format("staging:/%s", jar.getName());
-  }
-
-  private byte[] archiveToBytes(JavaArchive a) throws IOException {
-    ByteArrayOutputStream bout = new ByteArrayOutputStream();
-    a.as(ZipExporter.class)
-        .exportTo(bout);
-    byte[] byteArray = bout.toByteArray();
-    bout.close();
-    return byteArray;
+  public void deploy(LimbusStagingDeployment deployment) throws LimbusException, IOException {
+    requireNonNull(deployment, "Deploytment must not be null!");
+    LimbusEngine limbusEngine = system.getComponent(LimbusEngine.class);
+    limbusEngine.deployPlugin(deployment.getClasspath(), deployment.getPermissions());
   }
 
   /**
@@ -206,59 +123,58 @@ public final class LimbusStage {
     }
   }
 
-  void setDeployment(JavaArchive deployment) {
-    this.deployment = deployment;
-  }
-
-  JavaArchive getDeployment() {
-    return deployment;
-  }
-
-  void setDependencies(List<JavaArchive> dependencies) {
-    this.dependencies = dependencies;
-  }
-
-  void setSystem(LimbusSystem system) {
-    this.system = system;
-  }
-
-  void setDeployName(String deployName) {
-    this.deployName = deployName;
+  /**
+   * Adds the dependencies defined in the project's POM to the deployment. <b>Only call this method if you have
+   * dependencies defined in the project's POM.</b>
+   *
+   * @return Return this instance for method chaining.
+   */
+  public LimbusStage fromProjectDependencies() {
+    //@formatter:off
+    this.dependencies = Maven.configureResolver()
+                                    .withClassPathResolution(true)
+                                    .loadPomFromFile("pom.xml")
+                                    .importCompileAndRuntimeDependencies()
+                                    .resolve()
+                                    .withTransitivity()
+                                    .asList(JavaArchive.class);
+    //@formatter:on
+    return this;
   }
 
   /**
-   * Adds the specified permission to the deployment's classpath.
+   * Use this method to generate a dump of all dependencies of the current deployment.
    *
-   * <p>
-   * <b>This method only has an effect if the {@link LimbusStage} is started after calling this method.
-   * </p>
-   *
-   * @param permission
-   *        The permission to add.
+   * @return Returns this object for method chainging.
    */
-  public void addPermission(Permission permission) {
-    this.permissions.add(permission);
+  public LimbusStage dumpDeploymentDependencies() {
+    if (dependencies.isEmpty()) {
+      System.out.println("No deployment dependencies are defined yet.");
+    } else {
+      System.out.println("Printing out the deployment dependencies:");
+      for (JavaArchive d : dependencies) {
+        System.out.printf("- Dependency %s\n", d.getName());
+      }
+    }
+    return this;
   }
 
   /**
-   * Adds the specified set of {@link Permission}s to the deployment's classpath. If permissions were already defines,
-   * the specified set is added on top.
-   * <p>
-   * <b>This method only has an effect if the {@link LimbusStage} is started after calling this method.
-   * </p>
+   * Use this method to generate a dump of all files contained in the deployment's dependencies. The content of each
+   * archive that is deployed to the Limbus Engine as a dependency of the plugin deployment is printed.
    *
-   * @param permissions
-   *        The set of permissions to add.
+   * @return Returns this object for method chainging.
    */
-  public void addAllPermission(Set<Permission> permissions) {
-    this.permissions.addAll(permissions);
+  public LimbusStage dumpDeploymentDependenciesContent() {
+    if (dependencies.isEmpty()) {
+      System.out.println("No deployment dependencies are defined yet.");
+    } else {
+      System.out.println("Printing out the deployment dependencies archive content:");
+      for (JavaArchive d : dependencies) {
+        System.out.printf("Printing out the archive content of: %s\n", d.getName());
+        d.writeTo(System.out, Formatters.VERBOSE);
+      }
+    }
+    return this;
   }
-
-  /**
-   * @return Returns a new list of archives that are currently part of the deployment.
-   */
-  public List<JavaArchive> getDependencies() {
-    return new LinkedList<>(dependencies);
-  }
-
 }
